% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aooFixedRotationo}
\alias{aooFixedRotationo}
\title{BETA Area of Occupancy (AOO), from systematic grid rotation and shifting}
\usage{
aooFixedRotationo(
  thepoints,
  cellsize = 2000,
  it = 1296,
  returnV = "S",
  rotation = TRUE
)
}
\arguments{
\item{thepoints}{dataframe of points in metres i.e. c(x,y)}

\item{cellsize}{width of cell in metres (default 2000 m)}

\item{it}{the number of iterations you wish it to run, (default 1296)}

\item{returnV, }{switches to return different sets of results: \cr  \cr
S = Simple, returns just the minimum AOO area in km2, (DEFAULT)  \cr
E = Expended simple, returns the solution for the smallest AOO  as list of (area,number of cells, rotation (degrees), shift in x direction, shift in y direction)  \cr
ALL = returns a dataframe of all of the results from all the trials with (number of cells, rotation (0 in this case), shift in x direction (metres), shift in y direction (metres))  \cr
SF = returns a polygon simple feature for mapping, plotting in ggplot/plot or to export to a GIS format.}

\item{rotation}{allow rotation of grids? (default = TRUE). If rotations are trigger selected iterations are shared 50:50 rotation:shift(both in x and y direction)}
}
\value{
dependent on switch, default is area in km2
}
\description{
Calculates the Area of Occupancy AOO (smallest) by rotating the grid and shifting in x and y direction.
On a very few occasions the minimum solution will not always be achieved, but this solution is quick and consistent (not driven by the number of points).
If your species is near a threshold you may want to increase the number of iterations.
Please cite if using this algorithm: Moat, J., Bachman, S. P., Field, R., & Boyd, D. S. (2018). Refining area of occupancy to address the modifiable areal unit problem in ecology and conservation. Conservation biology, 32(6), 12781289. \cr
Works the same as aooFixedRotation, but much faster. In BETA until fully tested
On a very few occasions the minimum solution will not always be achieved but it is quick and consistent (not driven by the number of points).
If your species is near a threshold you may want to increase the number of iterations.
}
\seealso{
\code{\link{ratingAoo}} for AOO Ratings

\code{\link{aooFixedGrid}} for fixed grid optimal method

\code{\link{aooFixedRotation}} for original method method

\code{\link{aoo}} for simple AOO method

\code{\link{buildCellPolys_rxy}} for building grid polygons from points, rotation and shift
}
\author{
Justin Moat. J.Moat@kew.org
}
