% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildCellPolys_rxy}
\alias{buildCellPolys_rxy}
\title{Build simple feature polygons from point data, rotation and shift in x and y direction}
\usage{
buildCellPolys_rxy(thepoints, cellsize, rot, shiftx, shifty)
}
\arguments{
\item{thepoints}{set of points in metres i.e. c(x,y)}

\item{cellsize}{size of cell (length) in metres}

\item{rot}{rotation of the grid in radian}

\item{shiftx}{shift in the x direction in metres}

\item{shifty}{shift in the y direction in metres}
}
\value{
Simple Feature of polygons
}
\description{
Builds cell polygons (as simple features) from points and rotation, shift in X and y returns polygons for ggplot2 and mapping.
Generally used to plot data from AOO calculations.
}
\author{
Justin Moat. J.Moat@kew.org
}
