% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_occs.R
\name{check_occs}
\alias{check_occs}
\title{Checks and flags GBIF occurrence data based on common coordinate issues}
\usage{
check_occs(gbif_occs, native_ranges = NULL, buffer = 1000)
}
\arguments{
\item{gbif_occs}{(data frame) GBIF occurrence file in DWCA format}

\item{native_ranges}{(data frame) Optional native ranges from `get_native_range()`}

\item{buffer}{(integer) distance in meters of a buffer to account for coarse WGSRPD polygons.
Buffer is drawn around WGSRPD polygons and points within buffer are included in native range test.}
}
\value{
A list containing:
  - checked_data: original data with error flags
  - summary: summary of error counts
}
\description{
Checks and flags GBIF occurrence data based on common coordinate issues
}
