% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_tax_auth}
\alias{check_tax_auth}
\title{Resolves taxonomicAuthor field when there are varying taxonomic fields returned from GBIF query}
\usage{
check_tax_auth(df)
}
\arguments{
\item{df}{dataframe with gbif name search output}
}
\value{
returns GBIF taxonomic information with reformatted taxonomicAuthor
}
\description{
Resolves taxonomicAuthor field when there are varying taxonomic fields returned from GBIF query
}
