% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_occs.R
\name{clean_occs}
\alias{clean_occs}
\title{Cleans a GBIF occurrence download using CoordinateCleaner and native
range filter.}
\usage{
clean_occs(gbif_occs, native_ranges = NULL)
}
\arguments{
\item{gbif_occs}{(data frame) GBIF occurrence file in DWCA format. If [`get_gbif_occs()`] was used
to get GBIF occurrences file, note that the output is a list, so you need to e.g. my_occs$points}

\item{native_ranges}{(data frame) Native ranges prepared using [`get_native_range()`]}
}
\value{
a dataframe with cleaned occurrence data
}
\description{
Cleans a GBIF occurrence download using CoordinateCleaner and native
range filter.
}
