% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constructPolygon}
\alias{constructPolygon}
\title{Construct a polygon from vertices.}
\usage{
constructPolygon(x, y, crs)
}
\arguments{
\item{x}{East to West coordinate in metres}

\item{y}{South to North coordinate in metres}

\item{crs}{coordinate reference system}
}
\description{
Accepts an x and a y vector to define the vertices of
the polygon, to make it easier
}
