% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_species_map.R
\name{create_species_map}
\alias{create_species_map}
\title{Display a leaflet map to show the distribution of occurrence data for each species}
\usage{
create_species_map(data, species_range = NULL)
}
\arguments{
\item{data}{A data frame containing occurrence data}

\item{native_ranges}{A data frame containing occurrence data}
}
\value{
Interactive leaflet map
}
\description{
Display a leaflet map to show the distribution of occurrence data for each species
}
