% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_status_gauge.R
\name{create_status_gauge}
\alias{create_status_gauge}
\title{Generate metrics from a clean point occurrence file}
\usage{
create_status_gauge(species_data)
}
\arguments{
\item{species_data}{(df) results from make_metrics() function}
}
\value{
Returns a ggplot to visualise the Least Concern results
}
\description{
Generate metrics from a clean point occurrence file
}
