% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_occs.R
\name{flag_occs}
\alias{flag_occs}
\title{Flags GBIF occurrence data based on common coordinate issues}
\usage{
flag_occs(gbif_occs, native_ranges = NULL)
}
\arguments{
\item{gbif_occs}{(data frame) GBIF occurrence file in DWCA format}

\item{native_ranges}{(data frame) Optional native ranges from `get_native_range()`}
}
\value{
A list containing:
  - flagged_data: original data with error flags
  - summary: summary of error counts
}
\description{
Flags GBIF occurrence data based on common coordinate issues
}
