% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iso_occs.R
\name{get_iso_occs}
\alias{get_iso_occs}
\title{Get ISO two-digit code from occurrence lat longs}
\usage{
get_iso_occs(sis_points)
}
\arguments{
\item{sis_points}{(dataframe) SIS points file}
}
\value{
Occurrence dataframe with ISO two-digit codes added
}
\description{
Get ISO two-digit code from occurrence lat longs
}
