% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occs_range.R
\name{get_occs_range}
\alias{get_occs_range}
\title{Get ISO two-digit code from occurrence lat longs}
\usage{
get_occs_range(sis_points)
}
\arguments{
\item{sis_points}{(dataframe) SIS points file}
}
\value{
Occurrence dataframe with ISO two-digit codes added and internal id
}
\description{
Get ISO two-digit code from occurrence lat longs
}
