% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iucn_taxonomy}
\alias{iucn_taxonomy}
\title{'iucn_taxonomy'}
\format{
A data frame with 3942 rows and 5 variables:
\describe{
  \item{family}{chr}
  \item{kingdom}{chr}
  \item{phylum}{chr}
  \item{classname}{chr}
  \item{ordername}{chr}
}
}
\usage{
iucn_taxonomy
}
\description{
A dataset used internally by the LCr package.
Provides upper level taxonomic ranks used by IUCN Red List
}
\keyword{internal}
