% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lifeform_mapping_POWO_RL}
\alias{lifeform_mapping_POWO_RL}
\title{lifeform_mapping_POWO_RL}
\format{
A data frame with 448 rows and 3 variables:
\describe{
  \item{lifeform_description}{chr}
  \item{description}{chr}
  \item{code}{chr}
}
}
\usage{
lifeform_mapping_POWO_RL
}
\description{
A dataset used internally by the LCr package.
Maps POWO lifeform codes to IUCN life form codes
}
\keyword{internal}
