% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ll2cart}
\alias{ll2cart}
\title{Geographic coordinates to cartesian (x,y,z)}
\usage{
ll2cart(latr, longr)
}
\arguments{
\item{latr}{latitude point in radians}

\item{longr}{longtitude point in radians}
}
\value{
dataframe of x,y,z
}
\description{
Calculates the Cartesian coordinates (x,y,z) from lat long in radians. Used as part of the projection wizard.
}
\author{
Justin Moat. J.Moat@kew.org
}
