% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{llCheck}
\alias{llCheck}
\title{Check point data and latitude and longitude are sensible to work with}
\usage{
llCheck(thepoints)
}
\arguments{
\item{thepoints}{set of points as a dataframe with latitude and longitude}
}
\value{
nothing or warning or error
}
\description{
Checked the dataframe for NA, latitude below -90 or above 90, longitude below -180 and above 180, also warns if it finds whole numbers or 0
}
\author{
Justin Moat. J.Moat@kew.org
}
