% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longestAxis}
\alias{longestAxis}
\title{Longest distance from a set of points}
\usage{
longestAxis(thepoints, returnV = "S")
}
\arguments{
\item{thepoints}{dataframe of points of x,y}

\item{returnV, }{two switches either S for simply the distance or P for a dataframe of the two furthest points}
}
\value{
distance in metres or two points for the longest distance
}
\description{
Calculates the longest distances from a set of points
}
\note{
Useful as a scale for cellsize and location buffers, Willis et al 2003 suggest 1/10 of this for cellsize for AOO calculations as does Rivers et al (2010) for buffer distance for sub-population or location calculations.
}
\author{
Justin Moat. J.Moat@kew.org
}
