% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_biorealms.R
\name{make_biorealms}
\alias{make_biorealms}
\title{Generate a list of biorealms for a species}
\usage{
make_biorealms(native_ranges)
}
\arguments{
\item{TDWG_realms}{(dataframe) TDWG level 3 to realm converter}
}
\value{
concatenated list of biorealms in SIS connect format
}
\description{
Generate a list of biorealms for a species
}
