% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_elevation.R
\name{make_elevation}
\alias{make_elevation}
\title{Generate the elevation upper and lower bounds for the allfields.csv file}
\usage{
make_elevation(occs)
}
\arguments{
\item{occs}{}
}
\value{
Returns upper and lower elevation based on elevation field in GBIF occurrence file.
}
\description{
Generate the elevation upper and lower bounds for the allfields.csv file
}
