% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eoo_aoo.R
\name{make_eoo_aoo}
\alias{make_eoo_aoo}
\title{Generate the eoo and aoo values for the allfields.csv file}
\usage{
make_eoo_aoo(occs)
}
\arguments{
\item{occs}{}
}
\value{
Returns upper and lower elevation based on elevation field in GBIF occurrence file.
}
\description{
Generate the eoo and aoo values for the allfields.csv file
}
