% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_LC_points.R
\name{make_LC_points}
\alias{make_LC_points}
\title{Combined function to get clean SIS points file using name identifiers}
\usage{
make_LC_points(
  keys_df,
  first_name = "",
  second_name = "",
  institution = "",
  range_check = TRUE
)
}
\arguments{
\item{first_name}{(character) First name of assessor}

\item{second_name}{(character) Second name of assessor}

\item{institution}{(character) Name of institution or affiliation}

\item{range_check}{(boolean) TRUE if you want to carry out occurrence cleaning with native range filter}

\item{wcvp_ipni_id}{(character) WCVP identifier}

\item{gbif_id}{(character) GBIF_usageKey}
}
\value{
(list) Includes the GBIF citation, and "points" the cleaned SIS compatible point file
}
\description{
Combined function to get clean SIS points file using name identifiers
}
\details{
Designed for batch processing using keys derived from [`get_name_keys()`]
}
