% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sis_occs.R
\name{make_sis_occs}
\alias{make_sis_occs}
\title{Convert cleaned gbif DWCA structure to IUCN SIS format}
\usage{
make_sis_occs(occs_clean, first_name = "", second_name = "", institution = "")
}
\arguments{
\item{occs_clean}{(character) Cleaned GBIF occurrence file in DWCA format. See
occs_clean() function.}

\item{first_name}{(character) First name of the assessor}

\item{second_name}{(character) Second name of the assessor}

\item{institution}{(character) Institution/affiliation of the assessor}
}
\value{
An dataframe of occurence points aligned with IUCN Red List spatial data standards
}
\description{
Convert cleaned gbif DWCA structure to IUCN SIS format
}
