% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_zip.R
\name{make_zip}
\alias{make_zip}
\title{Generate the SIS connect zip file}
\usage{
make_zip(csv_list = list(sis_files), zip_filename = "SIS_files.zip")
}
\arguments{
\item{csv_list}{(list) List of SIS csv files you want to zip}

\item{zip_filename}{(character) Defaults to "singlezip.zip"}
}
\value{
Returns an SIS compliant zip file
}
\description{
Generate the SIS connect zip file
}
