% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_species_batch.R
\name{map_species_batch}
\alias{map_species_batch}
\title{Display a leaflet map to show the distribution of occurrence data for each species
Batch version that iterates through species}
\usage{
map_species_batch(
  data,
  species_range = NULL,
  show_flags = TRUE,
  save_map = TRUE
)
}
\arguments{
\item{data}{A data frame containing occurrence data after cleaning flags added}

\item{species_range}{A data frame containing occurrence data show_flags}

\item{show_flags}{Set to false if you just want to see the points map without flags}

\item{save_map}{Set to true if you want to save the maps as html files in the working directory}
}
\value{
Interactive leaflet map
}
\description{
Display a leaflet map to show the distribution of occurrence data for each species
Batch version that iterates through species
}
