% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_search_gbif.R
\name{name_search_gbif}
\alias{name_search_gbif}
\title{Query taxon name against GBIF backbone}
\usage{
name_search_gbif(name, species_rank = TRUE, match = "single")
}
\arguments{
\item{name}{(character) A taxon name}

\item{species_rank}{(character) Default TRUE returns matches at species
level, otherwise FALSE returns any taxonomic rank}

\item{match}{(character) Controls the output of the search. Use `single` to
force a single match result that has the highest confidence or `any` to return
all possible matches sorted in order of confidence}
}
\value{
Returns a data frame with initial search term and matching name(s)
}
\description{
Queries a taxon name against the GBIF names backbone. Simple wrapper around
the rgbif `name_backbone_verbose` function.
}
\examples{
name_search_gbif("Poa annua L.")
}
