% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_search_powo.R
\name{name_search_powo}
\alias{name_search_powo}
\title{Query taxon name against POWO (Plants of the World Online) backbone}
\usage{
name_search_powo(names_df, name_col, powo_tax_stat = "any")
}
\arguments{
\item{powo_tax_stat}{(character) Default `any` provides all taxonomic
matches, otherwise `accepted` returns only accepted names according to POWO}

\item{name_df}{(data frame) Taxon name(s)}
}
\value{
Returns a data frame with initial search term and matching name(s)
}
\description{
Queries a taxon name against the POWO names backbone. Simple wrapper around
the rWCVP `wcvp_match_names` function.
}
\examples{
name_search_gbif("Poa annua L.")
}
