% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_ref.R
\name{occs_gbif_ref}
\alias{occs_gbif_ref}
\title{Generate GBIF download citation according to IUCN format}
\usage{
occs_gbif_ref(gbif_meta)
}
\arguments{
\item{gbif_meta}{(character) GBIF download metadata}
}
\value{
Returns a data frame with IUCN SIS Connect reference
}
\description{
Generate GBIF download citation according to IUCN format
}
