% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powo_lookup.R
\name{powo_lookup}
\alias{powo_lookup}
\title{Lookup name in POWO (Plants of the World Online) using WCVP ID}
\usage{
powo_lookup(wcvp_ipni_id, distribution = FALSE)
}
\arguments{
\item{wcvp_ipni_id}{(character) WCVP identifier}

\item{distribution}{(character) Default `FALSE`returns taxon information.
Use `TRUE` to return geographic range information}
}
\value{
Returns a list with POWO data
}
\description{
Lookup name in POWO (Plants of the World Online) using WCVP ID
}
