% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{powo_range}
\alias{powo_range}
\title{Return native range of an accepted plant species from POWO (Plants of the
World Online) using WCVP identifier}
\usage{
powo_range(wcvp_ipni_id)
}
\arguments{
\item{wcvp_ipni_id}{(character) WCVP identifier}
}
\value{
(dataframe) Native range codes according to World Geographic Scheme for Recordings Plant Distributions (WGSRPD)
}
\description{
Return native range of an accepted plant species from POWO (Plants of the
World Online) using WCVP identifier
}
