% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powo_ref.R
\name{powo_ref}
\alias{powo_ref}
\title{Generate POWO download citation according to IUCN format}
\usage{
powo_ref()
}
\value{
Returns a dataframe with IUCN SIS Connect reference
}
\description{
This function uses queries POWO to get the native range
}
