% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powo_text.R
\name{powo_text}
\alias{powo_text}
\title{Lookup distribution, habit and climate text in POWO (Plants of the World Online) using WCVP ID}
\usage{
powo_text(wcvp_ipni_id, occs, unique_id)
}
\arguments{
\item{wcvp_ipni_id}{(character) WCVP identifier}

\item{occs}{(df) Occurrence data with elevation information}
}
\value{
Returns a list with POWO data
}
\description{
Lookup distribution, habit and climate text in POWO (Plants of the World Online) using WCVP ID
}
