% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pro2sph}
\alias{pro2sph}
\title{Cartesian coordinate projection}
\usage{
pro2sph(x, y, z)
}
\arguments{
\item{x}{East to West coordinate in metres}

\item{y}{South to North coordinate in metres}

\item{z}{height coordinate in metres}
}
\value{
x,y,z
}
\description{
Used as part of the projection wizard, calculates Cartesian (x,y,z), projected from the centre of the sphere to the earth surface, returns cartesian coordinates (x,y,z)
}
\note{
http://stackoverflow.com/questions/9604132/how-to-project-a-point-on-to-a-sphere
}
\author{
Justin Moat. J.Moat@kew.org
}
