% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ratingAoo}
\alias{ratingAoo}
\title{IUCN rating based on AOO Area}
\usage{
ratingAoo(AOOArea, abb = TRUE)
}
\arguments{
\item{AOOArea}{Area in km2}

\item{abb}{abbreviation TRUE or FALSE , TRUE = 2 letter code, FALSE = full text (default = TRUE)}
}
\value{
Text one of CR, EN, VU, NT, LC or Critically Endangered, Endangered, Vulnerable, Near Threatened, Least Concern
}
\description{
Calculates IUCN rating based on Area of occupancy (AOO) in km2
}
\author{
Justin Moat. J.Moat@kew.org
}
