% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ratingEoo}
\alias{ratingEoo}
\title{IUCN rating based on EOO Area}
\usage{
ratingEoo(EOOArea, abb = TRUE)
}
\arguments{
\item{EOOArea}{Area in km2}

\item{abb}{abbreviation TRUE or FALSE , TRUE = 2 letter code, FALSE = full text (see value), default = TRUE}
}
\value{
Text
one of CR, EN, VU, NT, LC or Critically Endangered, Endangered, Vulnerable, Near Threatened, Least Concern
}
\description{
Calculates IUCN rating based on Extent of Occurrence (EOO) Area in km2
}
\details{
Calculates IUCN rating on EOO
}
\seealso{
\code{\link{eoo}} for EOO calculations
}
\author{
Justin Moat. J.Moat@kew.org
}
