% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ratingPop}
\alias{ratingPop}
\title{IUCN rating based on population reduction}
\usage{
ratingPop(pReduction, subCr = 2, abb = TRUE)
}
\arguments{
\item{pReduction}{reduction as a percentage}

\item{subCr}{sub Criteria category: 1 or 2 or 3 or 4 (2 default)}

\item{abb}{abbreviation TRUE or FALSE , TRUE = 2 letter code, FALSE = full text (see value), default = TRUE}
}
\value{
Text
one of CR, EN, VU, NT, LC or Critically Endangered, Endangered, Vulnerable, Near Threatened, Least Concern
}
\description{
Calculates IUCN rating based on based on population reduction as a percentage
}
\details{
Calculates IUCN rating based on population reduction
}
\author{
Justin Moat. J.Moat@kew.org
}
