% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_status.R, R/utils.R
\name{reorder_status}
\alias{reorder_status}
\title{Reorder taxonomic status to give priority to accepted name}
\usage{
reorder_status(df)

reorder_status(df)
}
\arguments{
\item{df}{(character) data frame with names}
}
\value{
Returns a data frame with taxonomic status sorted

Returns a data frame with taxonomic status sorted
}
\description{
arranges order of taxonomic status from GBIF query

arranges order of taxonomic status from GBIF query
}
