% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotateP}
\alias{rotateP}
\title{Rotates a set of points}
\usage{
rotateP(thepoints, angle)
}
\arguments{
\item{thepoints}{set of points in metres i.e. c(x,y)}

\item{angle}{in radians}
}
\value{
dataframe of points
}
\description{
Rotates a set of point by an angle in radians. Used as part of the AOO rotation calculations.
}
\author{
Justin Moat. J.Moat@kew.org
}
