% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_name_powo.R
\name{search_name_powo}
\alias{search_name_powo}
\title{Query taxon name against POWO (Plants of the World Online) backbone}
\usage{
search_name_powo(df, name_column, author_col = NULL, names)
}
\arguments{
\item{df}{(data frame) Taxon name(s)}

\item{name_column}{(string) Column that contains the name(s)}

\item{author_col}{(string) Column that contains the author name(s)}

\item{names}{(data frame) A data frame of taxonomic names from WCVP}
}
\value{
Returns a data frame with initial search term and matching name(s)
}
\description{
Queries a taxon name against the POWO names backbone. Simple wrapper around
the rWCVP `wcvp_match_names` function.
}
