% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simProjWiz}
\alias{simProjWiz}
\title{Simple Projection Wizard}
\usage{
simProjWiz(thepoints, thecentre, returnV = "S")
}
\arguments{
\item{thepoints}{set of points as a dataframe with latitude and longitude}

\item{thecentre}{one point i.e. c(lat,long), if not specified this will be calculated from the center of gravity of all points}

\item{returnV}{switches to return either  dataframe (x,y) or simple feature of points  \cr
S = simple, returns as dataframe of x,y \cr
SF = simple feature of points}
}
\value{
Defaults is a set of points in meters as a dataframe with projection details attributed (stored as crs to retrieve attr(myprojectedpoints,'crs'))
}
\description{
Projects any set of latitude and longitude points to a "suitable" area projection, based on thieir "true centre of gravity".
Data is expected as lat long in decimal degrees and returned in metres.
Input data is checked to make sure it’s sensible before projection (i.e. lat and longs on the earth no null or NA values)
}
\author{
Justin Moat. J.Moat@kew.org
}
