% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_countries.R
\name{sis_countries}
\alias{sis_countries}
\title{Generate the countries.csv file}
\usage{
sis_countries(native_ranges, unique_id)
}
\arguments{
\item{native_ranges}{(character) native range list of WGSRPD level 3}

\item{unique_id}{(character) Unique identifier - default is the GBIF usage key}
}
\value{
Returns an SIS compliant csv file
}
\description{
Generate the countries.csv file
}
