% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_credits.R
\name{sis_credits}
\alias{sis_credits}
\title{Generate the credits.csv file}
\usage{
sis_credits(
  unique_id,
  first_name = "your first name",
  second_name = "your second name",
  email = "your email",
  affiliation = "your affiliation"
)
}
\arguments{
\item{unique_id}{(character) Unique identifier}

\item{first_name}{(character) First name of assessor}

\item{second_name}{(character) Second name of assessor}

\item{email}{(character) Email of assessor}

\item{affiliation}{(character) Name of institution or affiliation}
}
\value{
Returns an SIS compliant csv file
}
\description{
Generate the credits.csv file
}
