% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_habitats.R
\name{sis_habitats}
\alias{sis_habitats}
\title{Generate the habitats.csv file}
\usage{
sis_habitats(unique_id)
}
\arguments{
\item{unique_id}{(character) Unique identifier.}
}
\value{
Returns an SIS compliant csv file
}
\description{
Generate the habitats.csv file
}
