% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_map_single.R
\name{species_map_single}
\alias{species_map_single}
\title{Display a leaflet map to show the distribution of occurrence data for each species}
\usage{
species_map_single(data, species_range = NULL, show_flags = TRUE)
}
\arguments{
\item{data}{A data frame containing occurrence data}

\item{show_flags}{Set to false if you just want to see the points map without flags}

\item{native_ranges}{A data frame containing occurrence data show_flags}
}
\value{
Interactive leaflet map
}
\description{
Display a leaflet map to show the distribution of occurrence data for each species
}
