% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tdwg2iucn}
\alias{tdwg2iucn}
\title{'tdwg2iucn'}
\format{
A data frame with 866 rows and 5 variables:
\describe{
  \item{ID}{dbl}
  \item{Level.3.code}{chr}
  \item{TDWG_Name}{chr}
  \item{countryoccurrence.countryoccurrencesubfield.countryoccurrencename}{chr}
  \item{countryoccurrence.countryoccurrencesubfield.countryoccurrencelookup}{chr}
}
}
\usage{
tdwg2iucn
}
\description{
A dataset used internally by the LCr package.
Maps WGSRPD L3 TDWG values to IUCN land region values
}
\keyword{internal}
