% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tdwg_level3}
\alias{tdwg_level3}
\title{'tdwg_level3'}
\format{
A data frame with 369 rows and 5 variables:
\describe{
  \item{LEVEL3_NAM}{chr}
  \item{LEVEL3_COD}{chr}
  \item{LEVEL2_COD}{dbl}
  \item{LEVEL1_COD}{dbl}
  \item{geometry}{MULTIPOLYGON}
}
}
\usage{
tdwg_level3
}
\description{
A dataset used internally by the LCr package.
WGSRPD L3 TDWG values with geometry
}
\keyword{internal}
