% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TDWG_realms}
\alias{TDWG_realms}
\title{'TDWG_realms'}
\format{
A data frame with 394 rows and 4 variables:
\describe{
  \item{LEVEL3_NAM}{chr}
  \item{LEVEL3_COD}{chr}
  \item{FID_Ecoreg}{dbl}
  \item{REALM}{chr}
}
}
\usage{
TDWG_realms
}
\description{
A dataset used internally by the LCr package.
Describes biogeographic realms used by IUCN Red List
}
\keyword{internal}
