% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trueCOGll}
\alias{trueCOGll}
\title{True centre of gravity from a set of Lat longs}
\usage{
trueCOGll(thepoints)
}
\arguments{
\item{thepoints}{set of points c(lat,long)}
}
\value{
a point (lat,long) from centre
}
\description{
Calculates the "true" centre of gravity (weighted) from a set of lat longs, using cartesian geometry. Used as part of the projection wizard.
}
\author{
Justin Moat. J.Moat@kew.org
}
