% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchCon_new.R
\name{batchCon}
\alias{batchCon}
\title{Batch process, preliminary conservation assessments}
\usage{
batchCon(
  taxa,
  long,
  lat,
  project2gether = TRUE,
  cellsize = 2000,
  aooMin = FALSE,
  it = 1296,
  returnV = "S"
)
}
\arguments{
\item{taxa}{field which defines a list of species or taxa}

\item{long}{field which defines the longitude set of points}

\item{lat}{field which defines the latitude set of points}

\item{project2gether}{TRUE or FALSE, TRUE all data is projected together using the centre of all latitudes and longitudes. FALSE each species is projected separately. Default = TRUE}

\item{cellsize}{cell length in metres used to for AOO projection N.B. IUCN recommend 2000 m (default 2000)}

\item{aooMin}{calls the aooMin routines as well as simple aoo, be warned with lots of species and points this can take some time to run (default=FALSE)}

\item{it}{if aooMin=TRUE this determines the number of iterations it will run to find aooMin (default=1296)}

\item{returnV}{switches to return different sets of results: \cr
S = simple returns a dataframe of results = (taxa ,Number of points,EOO in km2, Simple AOO in km2,Minimum AOO, EOO category, AOOcategory, Cellwidth, projection parameters) \cr
SF = simple features dataframe will all results, taxa in taxon field, type=(EOO,AOO,points). NB all points will be projected together and aooMin is ignored}
}
\description{
Combines the main of routines in rCAT to process multiple species for AOO, EOO etc.
}
\details{
This function expects a list of taxa and latitudes and longitudes.ie\cr
\tabular{lll}{
species_w \tab 85.388000  \tab   84.33100\cr
species_w \tab -45.467000   \tab  88.41500\cr
species_w \tab -34.339000 \tab -149.52600\cr
species_x \tab -29.620000   \tab  79.11900\cr
species_x \tab 33.409000  \tab  -33.94700\cr
species_x \tab 64.692000 \tab -149.18900\cr
species_y \tab 2.308000 \tab -140.21900\cr
species_y \tab 41.452000  \tab -3.65600\cr
species_y \tab -30.475000 \tab -129.99600\cr}
etc
}
\note{
Has a switch to either project all data as a whole or each taxa separately. We would suggest you use this switch if data is all from a similar area (i.e. all from one continent/country/region)
}
\author{
Justin Moat. J.Moat@kew.org
}
