% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tax_auth.R
\name{check_tax_auth}
\alias{check_tax_auth}
\title{resolves taxonomicAuthor field when there are varying taxonimic fields returned from GBIF query}
\usage{
check_tax_auth(df)
}
\arguments{
\item{df}{}
}
\value{
returns GBIF taxonimic information with reformatted taxonomicAuthor
}
\description{
resolves taxonomicAuthor field when there are varying taxonimic fields returned from GBIF query
}
