% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_keys.R
\name{clean_keys}
\alias{clean_keys}
\title{Clean a keys output to ensure only accepted species from GBIF and WCVP are being analysed}
\usage{
clean_keys(keys)
}
\arguments{
\item{keys}{data frame of GBIF and WCVP identifier keys}
}
\value{
Returns a data frame with cleaned GBIF and WCVP keys
}
\description{
Clean a keys output to ensure only accepted species from GBIF and WCVP are being analysed
}
\details{
Problematic records e.g. not accepted in either GBIF or WCVP,
or not at species level, are removed. Duplicate records are reported, but not removed.
}
