% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_occs.R
\name{get_gbif_occs}
\alias{get_gbif_occs}
\title{Download georeferenced occurrence data using GBIF name identifier}
\usage{
get_gbif_occs(gbif_ids, keys_df = NULL)
}
\arguments{
\item{gbif_ids}{(character) GBIF identifiers}

\item{keys_df}{(character) WCVP identifiers.}
}
\value{
(list) 'Points' = the downloaded GBIF points, 'Citation' = the GBIF citation to be used in references
}
\description{
Download georeferenced occurrence data using GBIF name identifier
}
