% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif_occs.R
\name{get_gbif_occs}
\alias{get_gbif_occs}
\title{Download georeferenced occurrence data using GBIF name identifier}
\usage{
get_gbif_occs(keys_df)
}
\arguments{
\item{keys_df}{(data frame) Derived from [`get_name_keys()`] function. Must include at least GBIF_usageKey to obtain GBIF occurrences}
}
\value{
(list) 'Points' = the downloaded GBIF points, 'Citation' = the GBIF citation to be used in references
}
\description{
Download georeferenced occurrence data using GBIF name identifier
}
\details{
Initiates an occurrence download in DWCA format using [rgbif::occ_download()]. If `keys_df`contains multiple GBIF_usageKey values, they will be
combined in a single output file.
}
