% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_name_keys.R
\name{get_name_keys}
\alias{get_name_keys}
\title{Get accepted GBIF and POWO identifiers (keys) for taxon names to support data downloads}
\usage{
get_name_keys(names_df, name)
}
\arguments{
\item{names_df}{character) Dataframe with taxon names}

\item{name}{(character) Column for taxon names}
}
\value{
Returns a data frame with accepted GBIF and POWO identifiers
}
\description{
Get accepted GBIF and POWO identifiers (keys) for taxon names to support data downloads
}
