% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_native_range.R
\name{get_native_range}
\alias{get_native_range}
\title{Get native ranges for taxa}
\usage{
get_native_range(keys, name_col)
}
\arguments{
\item{keys}{(name_col) Column name for the identifier}
}
\value{
(data frame) A list of places where a taxon occurs.
}
\description{
Get native ranges for taxa
}
\details{
Currently one option to get native ranges from (Plants of the
World Online) using WCVP identifier. Other options to be added later e.g. GIFT
}
