% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCr-package.R
\docType{package}
\name{LCr-package}
\alias{LCr}
\alias{LCr-package}
\title{LCr: Generate minimal Least Concern Red List assessments}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generate Red List compliant spatial data and minimal Red List assessment documentation for Least Concern species
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stevenpbachman/LCr}
  \item \url{https://stevenpbachman.github.io/LCr/}
  \item Report bugs at \url{https://github.com/stevenpbachman/LCr/issues}
}

}
\author{
\strong{Maintainer}: Steven Bachman \email{s.bachman@kew.org} (\href{https://orcid.org/0000-0003-1085-6075}{ORCID})

Authors:
\itemize{
  \item Matilda Brown \email{m.brown2@kew.org} (\href{https://orcid.org/0000-0003-2536-8365}{ORCID})
}

}
\keyword{internal}
