% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_biorealms}
\alias{make_biorealms}
\title{Generate a list of biorealms for a species}
\usage{
make_biorealms(native_ranges)
}
\arguments{
\item{native_ranges}{(dataframe) native ranges}
}
\value{
concatenated list of biorealms in SIS connect format
}
\description{
Generate a list of biorealms for a species
}
