% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metrics.R
\name{make_metrics}
\alias{make_metrics}
\title{Generate metrics from a clean point occurrence file}
\usage{
make_metrics(occs_clean)
}
\arguments{
\item{occs_clean}{(character) Cleaned GBIF occurrence file in DWCA format. See
occs_clean() function.}
}
\value{
Returns a dataframe with species level metrics
}
\description{
Generate metrics from a clean point occurrence file
}
