% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sis_csvs.R
\name{make_sis_csvs}
\alias{make_sis_csvs}
\title{Generate all SIS connect csv files}
\usage{
make_sis_csvs(
  unique_id,
  wcvp_ipni_id,
  first_name,
  second_name,
  email,
  institution,
  gbif_ref = NULL,
  native_ranges,
  family,
  genus,
  species,
  taxonomicAuthority,
  kingdom = "plantae",
  powo_ref = FALSE
)
}
\arguments{
\item{first_name}{(character) First name of assessor}

\item{second_name}{(character) Second name of assessor}

\item{email}{(character) Email of assessor}

\item{institution}{(character) Name of institution or affiliation}

\item{gbif_ref}{(data frame) A GBIF download citation according to IUCN format.}

\item{native_ranges}{(data frame) Native ranges derived from [`get_native_range()`]}

\item{kingdom}{(character) Default is 'plantae', but can also be 'fungi'}

\item{keys}{(data frame) Derived from [`get_name_keys()`] function. Must include at least GBIF_usageKey to obtain GBIF occurrences}
}
\value{
Returns an SIS compliant zip file
}
\description{
Generate all SIS connect csv files
}
