% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_search_powo.R
\name{name_search_powo}
\alias{name_search_powo}
\title{Query taxon name against POWO (Plants of the World Online) backbone}
\usage{
name_search_powo(df, name_column, author_col = NULL)
}
\arguments{
\item{df}{(data frame) Taxon name(s)}

\item{name_column}{(string) Column that contains the name(s)}

\item{author_col}{(string) Column that contains the author name(s)}
}
\value{
Returns a data frame with initial search term and matching name(s)
}
\description{
Queries a taxon name against the POWO names backbone. Simple wrapper around
the rWCVP `wcvp_match_names` function.
}
