% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_directions.R, R/utils.R
\name{replace_cardinal_directions}
\alias{replace_cardinal_directions}
\title{Replace text from POWO so that it matches IUCN standards}
\usage{
replace_cardinal_directions(text)

replace_cardinal_directions(text)
}
\arguments{
\item{text}{(character) POWO distribution string}
}
\value{
IUCN standardised POWO range string

IUCN standardised POWO range string
}
\description{
Replace text from POWO so that it matches IUCN standards

Replace text from POWO so that it matches IUCN standards
}
