% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_name_gbif.R
\name{search_name_gbif}
\alias{search_name_gbif}
\title{Query taxon name against GBIF backbone}
\usage{
search_name_gbif(name, match = "single")
}
\arguments{
\item{name}{(character) A taxon name}

\item{match}{(character) Controls the output of the search. Use `single` to
force a single match result that has the highest confidence or `any` to return
all possible matches sorted in order of confidence}

\item{species_rank}{(character) Default TRUE returns matches at species
level, otherwise FALSE returns any taxonomic rank}
}
\value{
Returns a data frame with initial search term and matching name(s)
}
\description{
Queries a taxon name against the GBIF names backbone. Simple wrapper around
the rgbif `name_backbone_verbose` function.
}
\examples{
search_name_gbif("Poa annua L.")
}
