% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_allfields.R
\name{sis_allfields}
\alias{sis_allfields}
\title{Generate the allfields.csv file}
\usage{
sis_allfields(wcvp_ipni_id)
}
\arguments{
\item{wcvp_ipni_id}{(character) WCVP identifier.}
}
\value{
Returns an SIS compliant data frame
}
\description{
Generate the allfields.csv file
}
