% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_assessments.R
\name{sis_assessments}
\alias{sis_assessments}
\title{Generate the assessments.csv file}
\usage{
sis_assessments(wcvp_ipni_id)
}
\arguments{
\item{wcvp_ipni_id}{(character) WCVP identifier.}
}
\value{
Returns an SIS compliant data frame
}
\description{
Generate the assessments.csv file
}
