% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_plantspecific.R
\name{sis_plantspecific}
\alias{sis_plantspecific}
\title{Generate the plantspecific.csv file}
\usage{
sis_plantspecific(unique_id)
}
\arguments{
\item{unique_id}{(character) Unique identifier.}
}
\value{
An SIS compliant csv file
}
\description{
Generate the plantspecific.csv file
}
