% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_references.R
\name{sis_references}
\alias{sis_references}
\title{Generate the references.csv file}
\usage{
sis_references(unique_id, gbif_ref, powo_ref)
}
\arguments{
\item{unique_id}{(character) Unique identifier.}

\item{gbif_ref}{(data frame) A GBIF download citation according to IUCN format.}

\item{powo_ref}{(logical) Adds POWO reference if TRUE.}
}
\value{
Returns an SIS compliant data frame
}
\description{
Generate the references.csv file
}
