% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_taxonomy.R
\name{sis_taxonomy}
\alias{sis_taxonomy}
\title{Generate the taxonomy.csv file}
\usage{
sis_taxonomy(unique_id, family, genus, species, taxonomicAuthority)
}
\arguments{
\item{unique_id}{(character) Unique identifier.}
}
\value{
Returns an SIS compliant csv file
}
\description{
Generate the taxonomy.csv file
}
